Junk'd_Project/Python_AI_API/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
tensorflow==2.13.0
flask
flask-cors
pillow
numpy


# Optional: For actual AI model implementation
# tensorflow==2.13.0
# torch==2.0.1
# torchvision==0.15.2
# opencv-python==4.8.1.78
# scikit-learn==1.3.0
# pandas==2.0.3
# matplotlib==3.7.2